http://download.virtualbox.org/virtualbox/2.1.0/VirtualBox-2.1.0-OSE.tar.bz2

Merge commit 'e8553686435257137b6a16e65f6bb1ac79edb205' of git://github.com/larsr/VirtualBox-OSE into VirtualBox/2.1/tar

Conflicts:
	Config.kmk
	Makefile.kmk
	debian/LocalConfig.kmk
	debian/changelog
	debian/lintian-override
	debian/postinst
	debian/rules
	include/VBox/VBoxGuest.h
	src/VBox/Additions/WINNT/VBoxGuest/NTLegacy.cpp
	src/VBox/Additions/WINNT/VBoxGuest/VBoxGuestPnP.cpp
	src/VBox/Additions/WINNT/VBoxService/VBoxVMInfo.cpp
	src/VBox/Additions/WINNT/VBoxTray/VBoxClipboard.cpp
	src/VBox/Additions/common/VBoxControl/VBoxControl.cpp
	src/VBox/Additions/common/VBoxGuestLib/VBoxGuestR3LibGuestProp.cpp
	src/VBox/Additions/x11/xclient/clipboard.cpp
	src/VBox/Debugger/DBGCBuiltInSymbols.cpp
	src/VBox/Debugger/DBGCCmdHlp.cpp
	src/VBox/Debugger/DBGCCmdWorkers.cpp
	src/VBox/Debugger/DBGCCommands.cpp
	src/VBox/Debugger/DBGCEmulateCodeView.cpp
	src/VBox/Debugger/DBGCInternal.h
	src/VBox/Debugger/DBGCOps.cpp
	src/VBox/Debugger/DBGCTcp.cpp
	src/VBox/Debugger/DBGConsole.cpp
	src/VBox/Debugger/DBGPlugInCommonELF.cpp
	src/VBox/Debugger/DBGPlugInCommonELF.h
	src/VBox/Debugger/DBGPlugInSolaris.cpp
	src/VBox/Debugger/Makefile.kmk
	src/VBox/Debugger/VBoxDbg.cpp
	src/VBox/Debugger/VBoxDbgBase.cpp
	src/VBox/Debugger/VBoxDbgBase.h
	src/VBox/Debugger/VBoxDbgConsole.cpp
	src/VBox/Debugger/VBoxDbgConsole.h
	src/VBox/Debugger/VBoxDbgGui.cpp
	src/VBox/Debugger/VBoxDbgGui.h
	src/VBox/Debugger/VBoxDbgStats.cpp
	src/VBox/Debugger/VBoxDbgStats.h
	src/VBox/Debugger/testcase/tstDBGCParser.cpp
	src/VBox/Debugger/testcase/tstDBGCStubs.cpp
	src/VBox/Devices/Builtins.cpp
	src/VBox/Devices/PC/BIOS/biossums.c
	src/VBox/Devices/PC/DevPit-i8254.cpp
	src/VBox/Devices/PC/DrvACPI.cpp
	src/VBox/Devices/Storage/DrvVD.cpp
	src/VBox/Devices/Storage/RawHDDCore.cpp
	src/VBox/Devices/Storage/VBoxHDD-new.cpp
	src/VBox/Devices/Storage/VBoxHDD.cpp
	src/VBox/Devices/Storage/VDICore.h
	src/VBox/Devices/Storage/VHDHDDCore.cpp
	src/VBox/Devices/Storage/VmdkHDDCore.cpp
	src/VBox/Devices/testcase/tstDeviceStructSizeGC.cpp
	src/VBox/Devices/vl_vbox.h
	src/VBox/Frontends/VBoxManage/VBoxManage.cpp
	src/VBox/Frontends/VBoxSDL/Makefile.kmk
	src/VBox/Frontends/VirtualBox/Makefile.kmk
	src/VBox/Frontends/VirtualBox/src/VBoxProblemReporter.cpp
	src/VBox/Frontends/VirtualBox4/Makefile.kmk
	src/VBox/Frontends/VirtualBox4/include/VBoxSelectorWnd.h
	src/VBox/Frontends/VirtualBox4/nls/ApprovedLanguages.kmk
	src/VBox/Frontends/VirtualBox4/nls/VirtualBox_de.ts
	src/VBox/Frontends/VirtualBox4/nls/VirtualBox_fr.ts
	src/VBox/Frontends/VirtualBox4/src/VBoxConsoleView.cpp
	src/VBox/Frontends/VirtualBox4/src/VBoxConsoleWnd.cpp
	src/VBox/Frontends/VirtualBox4/src/VBoxDiskImageManagerDlg.cpp
	src/VBox/Frontends/VirtualBox4/src/VBoxGlobal.cpp
	src/VBox/Frontends/VirtualBox4/src/VBoxProblemReporter.cpp
	src/VBox/Frontends/VirtualBox4/src/VBoxSelectorWnd.cpp
	src/VBox/Frontends/VirtualBox4/src/main.cpp
	src/VBox/HostDrivers/Support/Makefile.kmk
	src/VBox/HostDrivers/Support/SUPDrv.c
	src/VBox/HostDrivers/Support/SUPR3HardenedMain.cpp
	src/VBox/HostDrivers/VBoxNetFlt/VBoxNetFltInternal.h
	src/VBox/HostDrivers/VBoxNetFlt/darwin/loadnetflt.sh
	src/VBox/HostDrivers/VBoxNetFlt/solaris/VBoxNetFlt-solaris.c
	src/VBox/HostServices/SharedClipboard/x11.cpp
	src/VBox/HostServices/SharedOpenGL/Makefile.kmk
	src/VBox/HostServices/SharedOpenGL/unpacker/unpack_visibleregion.c
	src/VBox/Installer/linux/Makefile.kmk
	src/VBox/Installer/linux/vboxnet.sh.in
	src/VBox/Installer/solaris/Makefile.kmk
	src/VBox/Installer/solaris/postinstall.sh
	src/VBox/Installer/solaris/smf-vboxwebsrv.sh
	src/VBox/Installer/solaris/vboxdrv.sh
	src/VBox/Installer/solaris/virtualbox-zoneaccess.xml
	src/VBox/Installer/solaris/zoneaccess.sh
	src/VBox/Main/ConsoleImpl.cpp
	src/VBox/Main/ConsoleImpl2.cpp
	src/VBox/Main/MachineImpl.cpp
	src/VBox/Main/Makefile.kmk
	src/VBox/Main/VirtualBoxImpl.cpp
	src/VBox/Main/idl/VirtualBox.xidl
	src/VBox/Main/include/ConsoleImpl.h
	src/VBox/Main/include/HardDiskImpl.h
	src/VBox/Main/include/MachineImpl.h
	src/VBox/Main/include/Performance.h
	src/VBox/Main/include/VirtualBoxImpl.h
	src/VBox/Main/testcase/Makefile.kmk
	src/VBox/Main/win/PerformanceWin.cpp
	src/VBox/Runtime/darwin/RTErrConvertFromDarwinCOM.cpp
	src/VBox/Runtime/darwin/RTErrConvertFromDarwinIO.cpp
	src/VBox/Runtime/darwin/RTErrConvertFromDarwinKern.cpp
	src/VBox/Runtime/r0drv/solaris/vbi/semevent-r0drv-solaris.c
	src/VBox/Runtime/r0drv/solaris/vbi/semeventmulti-r0drv-solaris.c
	src/VBox/Runtime/r0drv/solaris/vbi/semfastmutex-r0drv-solaris.c
	src/VBox/Runtime/r0drv/solaris/vbi/spinlock-r0drv-solaris.c
	src/VBox/VMM/PATM/PATM.cpp
	src/VBox/VMM/VMM.cpp
	src/VBox/VMM/VMMR0/HWACCMR0A.asm
	src/VBox/VMM/VMMR0/VMMR0.cpp

+++ include/iprt/path.h
@@ -152,7 +152,7 @@ RTDECL(int)  RTPathGetCurrent(char *pszPath, size_t cchPath);
  * @param   pszRealPath     Where to store the real path.
  * @param   cchRealPath     Size of the buffer.
  */
-RTDECL(int) RTPathReal(const char *pszPath, char *pszRealPath, unsigned cchRealPath);
+RTDECL(int) RTPathReal(const char *pszPath, char *pszRealPath, size_t cchRealPath);
 
 /**
  * Same as RTPathReal only the result is RTStrDup()'ed.
@@ -171,7 +171,7 @@ RTDECL(char *) RTPathRealDup(const char *pszPath);
  * @param   pszAbsPath      Where to store the absolute path.
  * @param   cchAbsPath      Size of the buffer.
  */
-RTDECL(int) RTPathAbs(const char *pszPath, char *pszAbsPath, unsigned cchAbsPath);
+RTDECL(int) RTPathAbs(const char *pszPath, char *pszAbsPath, size_t cchAbsPath);
 
 /**
  * Same as RTPathAbs only the result is RTStrDup()'ed.
@@ -195,7 +195,7 @@ RTDECL(char *) RTPathAbsDup(const char *pszPath);
  * @param   pszAbsPath      Where to store the absolute path.
  * @param   cchAbsPath      Size of the buffer.
  */
-RTDECL(int) RTPathAbsEx(const char *pszBase, const char *pszPath, char *pszAbsPath, unsigned cchAbsPath);
+RTDECL(int) RTPathAbsEx(const char *pszBase, const char *pszPath, char *pszAbsPath, size_t cchAbsPath);
 
 /**
  * Same as RTPathAbsEx only the result is RTStrDup()'ed.
@@ -233,12 +233,12 @@ RTDECL(void) RTPathStripTrailingSlash(char *pszPath);
 
 /**
  * Parses a path.
- * 
- * It figures the length of the directory component, the offset of 
+ *
+ * It figures the length of the directory component, the offset of
  * the file name and the location of the suffix dot.
  *
  * @returns The path length.
- * 
+ *
  * @param   pszPath     Path to find filename in.
  * @param   pcbDir      Where to put the length of the directory component.
  *                      If no directory, this will be 0. Optional.
@@ -248,7 +248,7 @@ RTDECL(void) RTPathStripTrailingSlash(char *pszPath);
  * @param   poffSuff    Where to store the suffix offset (the last dot).
  *                      If empty string or if it's ending with a slash this
  *                      will be set to -1. Optional.
- * @param   pfFlags     Where to set flags returning more information about 
+ * @param   pfFlags     Where to set flags returning more information about
  *                      the path. For the future. Optional.
  */
 RTDECL(size_t) RTPathParse(const char *pszPath, size_t *pcchDir, ssize_t *poffName, ssize_t *poffSuff);
@@ -348,7 +348,7 @@ RTDECL(bool) RTPathStartsWith(const char *pszPath, const char *pszParentPath);
  * @param   pszPath     Buffer where to store the path.
  * @param   cchPath     Buffer size in bytes.
  */
-RTDECL(int) RTPathProgram(char *pszPath, unsigned cchPath);
+RTDECL(int) RTPathProgram(char *pszPath, size_t cchPath);
 
 /**
  * Gets the user home directory.
@@ -357,7 +357,7 @@ RTDECL(int) RTPathProgram(char *pszPath, unsigned cchPath);
  * @param   pszPath     Buffer where to store the path.
  * @param   cchPath     Buffer size in bytes.
  */
-RTDECL(int) RTPathUserHome(char *pszPath, unsigned cchPath);
+RTDECL(int) RTPathUserHome(char *pszPath, size_t cchPath);
 
 /**
  * Gets the directory of shared libraries. This is not the same as
@@ -372,7 +372,7 @@ RTDECL(int) RTPathUserHome(char *pszPath, unsigned cchPath);
  * @param   pszPath     Buffer where to store the path.
  * @param   cchPath     Buffer size in bytes.
  */
-RTDECL(int) RTPathSharedLibs(char *pszPath, unsigned cchPath);
+RTDECL(int) RTPathSharedLibs(char *pszPath, size_t cchPath);
 
 /**
  * Gets the directory for architecture-independent application data, for
@@ -386,7 +386,7 @@ RTDECL(int) RTPathSharedLibs(char *pszPath, unsigned cchPath);
  * @param   pszPath     Buffer where to store the path.
  * @param   cchPath     Buffer size in bytes.
  */
-RTDECL(int) RTPathAppPrivateNoArch(char *pszPath, unsigned cchPath);
+RTDECL(int) RTPathAppPrivateNoArch(char *pszPath, size_t cchPath);
 
 /**
  * Gets the directory for architecture-dependent application data, for
@@ -400,7 +400,7 @@ RTDECL(int) RTPathAppPrivateNoArch(char *pszPath, unsigned cchPath);
  * @param   pszPath     Buffer where to store the path.
  * @param   cchPath     Buffer size in bytes.
  */
-RTDECL(int) RTPathAppPrivateArch(char *pszPath, unsigned cchPath);
+RTDECL(int) RTPathAppPrivateArch(char *pszPath, size_t cchPath);
 
 /**
  * Gets the directory for documentation.
@@ -413,7 +413,7 @@ RTDECL(int) RTPathAppPrivateArch(char *pszPath, unsigned cchPath);
  * @param   pszPath     Buffer where to store the path.
  * @param   cchPath     Buffer size in bytes.
  */
-RTDECL(int) RTPathAppDocs(char *pszPath, unsigned cchPath);
+RTDECL(int) RTPathAppDocs(char *pszPath, size_t cchPath);
 
 /**
  * Query information about a file system object.
